requirements.txt
Django==2.1
django-rest-auth==0.9.3
djangorestframework==3.8.2
Pillow==5.2.0
pytz==2018.5
six==1.11.0
